@@ -2,9 +2,16 @@ import Vapor
 import Fluent
 
 extension DeviceSession: ModelTokenAuthenticatable {
-    static let valueKey = \DeviceSession.$accessToken
-    static let userKey = \DeviceSession.$user
-
+    typealias User = App.User
+    
+    static var valueKey: KeyPath<DeviceSession, Field<String>> {
+        \DeviceSession.$accessToken
+    }
+    
+    static var userKey: KeyPath<DeviceSession, Parent<User>> {
+        \DeviceSession.$user
+    }
+    
     var isValid: Bool {
         true
     }
