@@ -2,8 +2,14 @@ import Vapor
 import Fluent
 
 extension User: ModelAuthenticatable {
-    static let usernameKey = \User.$username
-    static let passwordHashKey = \User.$passwordHash
+
+    static var usernameKey: KeyPath<User, Field<String>> {
+        \User.$username
+    }
+
+    static var passwordHashKey: KeyPath<User, Field<String>> {
+        \User.$passwordHash
+    }
 
     func verify(password: String) throws -> Bool {
         try Bcrypt.verify(password, created: self.passwordHash)
