@@ -3,7 +3,7 @@ import XCTVapor
 
 final class TestPushManager: PushSender {
     
-    func send(_ notification: Service.Notification, to device: DeviceInfo) {
+    func send(_ notification: CoreService.Notification, to device: DeviceInfo) {
         guard let deviceToken = device.token else { return print("Can't send push without device token.") }
         print("--- TEST '\(device.transport.rawValue.uppercased())' push '\(notification.event)' sent to device '\(deviceToken)' with source '\(notification.source)' and payload: \(String(describing: notification.payload))")
     }
@@ -15,21 +15,29 @@ final class TestWebSocketManager: WebSocketServer, WebSocketSender {
         print("Accepted client at address \(clientAddress)")
     }
     
-    func send(_ notification: Service.Notification, to session: DeviceSession) async throws -> Bool {
+    func send(_ notification: CoreService.Notification, to session: DeviceSession) async throws -> Bool {
         guard let channel = session.id?.uuidString else { throw ServiceError(.internalServerError) }
         print("--- Test Message '\(notification.event)' sent to channel '\(channel)' with source '\(notification.source)' and data: \(String(describing: notification.payload))")
         return true
     }
 }
 
-final class TestNotificationManager: Notificator {
+actor TestNotificationManager: Notificator {
     
-    var sentNotifications = [Service.Notification]()
+    private var sentNotifications = [CoreService.Notification]()
     
-    func notify(chat: Chat, about event: Service.Event, from user: User?, with payload: JSON?) async throws {
+    func getSentNotifications() -> [CoreService.Notification] {
+        sentNotifications
+    }
+    
+    func clearSentNotifications() {
+        sentNotifications.removeAll()
+    }
+    
+    func notify(chat: Chat, in repo: ChatsRepository, about event: CoreService.Event, from user: User?, with payload: JSON?) async throws {
         let source = user == nil ? "system" : "\(user!.id ?? 0)"
-        var notification = Service.Notification(event: event, source: source, payload: payload)
-        let relations = try await Service.shared.chats.repo.findRelations(of: chat.id!, isUserBlocked: false)
+        var notification = CoreService.Notification(event: event, source: source, payload: payload)
+        let relations = try await repo.findRelations(of: chat.id!, isUserBlocked: false)
         let allowed = relations.filter { !$0.isChatBlocked }
         for relation in allowed {
             notification.destination = "\(relation.user.id!)"
@@ -39,7 +47,7 @@ final class TestNotificationManager: Notificator {
     }
 }
 
-extension Service {
+extension CoreService {
     
     var testNotificator: TestNotificationManager {
         notificator as! TestNotificationManager
