@@ -1,17 +1,7 @@
 @testable import App
 import XCTVapor
 
-final class UserTests: XCTestCase {
-    
-    var app: Application!
-    
-    override func setUp() {
-        app = try! Application.testable()
-    }
-    
-    override func tearDown() {
-        app.shutdown()
-    }
+final class UserTests: AppTestCase {
     
     func testRegisterUser() throws {
         try app.test(.POST, "users", beforeRequest: { req in
@@ -28,11 +18,11 @@ final class UserTests: XCTestCase {
     }
     
     func testDeregisterUser() async throws {
-        try await seedCurrentUser()
+        try await service.seedCurrentUser()
         var tokenString = ""
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -43,29 +33,29 @@ final class UserTests: XCTestCase {
             XCTAssertNotNil(session)
             tokenString = session!.accessToken
         })
-        try app.test(.GET, "users/me",
-                     headers: .authWith(token: tokenString),
-                     afterResponse: { res in
+        try await asyncTest(.GET, "users/me",
+                            headers: .authWith(token: tokenString),
+                            afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
         })
-        try app.test(.DELETE, "users/me",
-                     headers: .authWith(token: tokenString),
-                     afterResponse: { res in
+        try await asyncTest(.DELETE, "users/me",
+                            headers: .authWith(token: tokenString),
+                            afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
         })
-        try app.test(.GET, "users/me",
-                     headers: .authWith(token: tokenString),
-                     afterResponse: { res in
+        try await asyncTest(.GET, "users/me",
+                            headers: .authWith(token: tokenString),
+                            afterResponse: { res in
             XCTAssertEqual(res.status, .unauthorized, res.body.string)
         })
     }
     
     func testLoginUser() async throws {
-        try await seedCurrentUser()
+        try await service.seedCurrentUser()
         var tokenString = ""
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -77,26 +67,26 @@ final class UserTests: XCTestCase {
             tokenString = session!.accessToken
         })
         
-        try app.test(.GET, "users/me",
-                     headers: .authWith(token: tokenString),
-                     afterResponse: { res in
+        try await asyncTest(.GET, "users/me",
+                            headers: .authWith(token: tokenString),
+                            afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
             let response = try res.content.decode(User.PrivateInfo.self)
             XCTAssertEqual(response.info.username, CurrentUser.username)
         })
         
-        try app.test(.GET, "users/me",
-                     headers: .authWith(token: "fake"),
-                     afterResponse: { res in
+        try await asyncTest(.GET, "users/me",
+                            headers: .authWith(token: "fake"),
+                            afterResponse: { res in
             XCTAssertEqual(res.status, .unauthorized, res.body.string)
         })
     }
     
     func testLoginUserFailure() async throws {
-        try await seedCurrentUser()
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: "123"),
-                     beforeRequest: { req in
+        try await service.seedCurrentUser()
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: "123"),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -106,11 +96,11 @@ final class UserTests: XCTestCase {
     }
     
     func testCurrentUser() async throws {
-        try await seedCurrentUser()
+        try await service.seedCurrentUser()
         var tokenString = ""
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -121,19 +111,19 @@ final class UserTests: XCTestCase {
             XCTAssertNotNil(session)
             tokenString = session!.accessToken
         })
-        try app.test(.GET, "users/me",
-                     headers: .authWith(token: tokenString),
-                     afterResponse: { res in
+        try await asyncTest(.GET, "users/me",
+                            headers: .authWith(token: tokenString),
+                            afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
         })
     }
     
     func testLogoutUser() async throws {
-        try await seedCurrentUser()
+        try await service.seedCurrentUser()
         var tokenString = ""
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -144,29 +134,29 @@ final class UserTests: XCTestCase {
             XCTAssertNotNil(session)
             tokenString = session!.accessToken
         })
-        try app.test(.GET, "users/me",
-                     headers: .authWith(token: tokenString),
-                     afterResponse: { res in
+        try await asyncTest(.GET, "users/me",
+                            headers: .authWith(token: tokenString),
+                            afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
         })
-        try app.test(.POST, "users/me/logout",
-                     headers: .authWith(token: tokenString),
-                     afterResponse: { res in
+        try await asyncTest(.POST, "users/me/logout",
+                            headers: .authWith(token: tokenString),
+                            afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
         })
-        try app.test(.GET, "users/me",
-                     headers: .authWith(token: tokenString),
-                     afterResponse: { res in
+        try await asyncTest(.GET, "users/me",
+                            headers: .authWith(token: tokenString),
+                            afterResponse: { res in
             XCTAssertEqual(res.status, .unauthorized, res.body.string)
         })
     }
     
     func testChangePassword() async throws {
-        try await seedCurrentUser()
+        try await service.seedCurrentUser()
         var tokenString = ""
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -177,18 +167,18 @@ final class UserTests: XCTestCase {
             XCTAssertNotNil(session)
             tokenString = session!.accessToken
         })
-        try app.test(.PUT, "users/me/changePassword",
-                     headers: .authWith(token: tokenString),
-                     beforeRequest: { req in
+        try await asyncTest(.PUT, "users/me/changePassword",
+                            headers: .authWith(token: tokenString),
+                            beforeRequest: { req in
             try req.content.encode(
                 ChangePasswordRequest(oldPassword: CurrentUser.password, newPassword: "12345678")
             )
         }, afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
         })
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: "12345678"),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: "12345678"),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -198,11 +188,11 @@ final class UserTests: XCTestCase {
     }
     
     func testTryChangePasswordWithIncorrectCurrentPassword() async throws {
-        try await seedCurrentUser()
+        try await service.seedCurrentUser()
         var tokenString = ""
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -213,9 +203,9 @@ final class UserTests: XCTestCase {
             XCTAssertNotNil(session)
             tokenString = session!.accessToken
         })
-        try app.test(.PUT, "users/me/changePassword",
-                     headers: .authWith(token: tokenString),
-                     beforeRequest: { req in
+        try await asyncTest(.PUT, "users/me/changePassword",
+                            headers: .authWith(token: tokenString),
+                            beforeRequest: { req in
             try req.content.encode(
                 ChangePasswordRequest(oldPassword: "87654321", newPassword: "12345678")
             )
@@ -225,18 +215,18 @@ final class UserTests: XCTestCase {
     }
     
     func testResetPassword() async throws {
-        try await seedCurrentUser()
-        try app.test(.POST, "users/resetPassword",
-                     beforeRequest: { req in
+        try await service.seedCurrentUser()
+        try await asyncTest(.POST, "users/resetPassword",
+                            beforeRequest: { req in
             try req.content.encode(
                 ResetPasswordRequest(userId: 1, newPassword: "12345678", accountKey: CurrentUser.accountKey)
             )
         }, afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
         })
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: "12345678"),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: "12345678"),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -246,9 +236,9 @@ final class UserTests: XCTestCase {
     }
     
     func testTryResetPasswordWithInvalidAccountKey() async throws {
-        try await seedCurrentUser()
-        try app.test(.POST, "users/resetPassword",
-                     beforeRequest: { req in
+        try await service.seedCurrentUser()
+        try await asyncTest(.POST, "users/resetPassword",
+                            beforeRequest: { req in
             try req.content.encode(
                 ResetPasswordRequest(userId: 1, newPassword: "12345678", accountKey: "")
             )
@@ -258,9 +248,9 @@ final class UserTests: XCTestCase {
     }
     
     func testTryResetPasswordWithAccountKeyNotSet() async throws {
-        try await seedCurrentUser(accountKey: nil)
-        try app.test(.POST, "users/resetPassword",
-                     beforeRequest: { req in
+        try await service.seedCurrentUser(accountKey: nil)
+        try await asyncTest(.POST, "users/resetPassword",
+                            beforeRequest: { req in
             try req.content.encode(
                 ResetPasswordRequest(userId: 1, newPassword: "12345678", accountKey: CurrentUser.accountKey)
             )
@@ -270,11 +260,11 @@ final class UserTests: XCTestCase {
     }
     
     func testSetAccountKey() async throws {
-        try await seedCurrentUser(accountKey: nil)
+        try await service.seedCurrentUser(accountKey: nil)
         var tokenString = ""
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -285,17 +275,17 @@ final class UserTests: XCTestCase {
             XCTAssertNotNil(session)
             tokenString = session!.accessToken
         })
-        try app.test(.PUT, "users/me/setAccountKey",
-                     headers: .authWith(token: tokenString),
-                     beforeRequest: { req in
+        try await asyncTest(.PUT, "users/me/setAccountKey",
+                            headers: .authWith(token: tokenString),
+                            beforeRequest: { req in
             try req.content.encode(
                 SetAccountKeyRequest(password: CurrentUser.password, accountKey: "\(CurrentUser.accountKey.reversed())")
             )
         }, afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
         })
-        try app.test(.POST, "users/resetPassword",
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/resetPassword",
+                            beforeRequest: { req in
             try req.content.encode(
                 ResetPasswordRequest(userId: 1, newPassword: "12345678", accountKey: "\(CurrentUser.accountKey.reversed())")
             )
@@ -305,11 +295,11 @@ final class UserTests: XCTestCase {
     }
     
     func testTrySetAccountKeyWithIncorrectCurrentPassword() async throws {
-        try await seedCurrentUser(accountKey: nil)
+        try await service.seedCurrentUser(accountKey: nil)
         var tokenString = ""
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -320,9 +310,9 @@ final class UserTests: XCTestCase {
             XCTAssertNotNil(session)
             tokenString = session!.accessToken
         })
-        try app.test(.PUT, "users/me/setAccountKey",
-                     headers: .authWith(token: tokenString),
-                     beforeRequest: { req in
+        try await asyncTest(.PUT, "users/me/setAccountKey",
+                            headers: .authWith(token: tokenString),
+                            beforeRequest: { req in
             try req.content.encode(
                 SetAccountKeyRequest(password: "123", accountKey: "\(CurrentUser.accountKey.reversed())")
             )
@@ -332,9 +322,9 @@ final class UserTests: XCTestCase {
     }
     
     func testUpdateUser() async throws {
-        try await seedCurrentUser()
+        try await service.seedCurrentUser()
         let about = UUID().uuidString
-        try app.test(.PUT, "users/me", headers: .none, beforeRequest: { req in
+        try await asyncTest(.PUT, "users/me", headers: .none, beforeRequest: { req in
             try req.content.encode([
                 "name": "Test Test",
                 "about": about
@@ -350,11 +340,11 @@ final class UserTests: XCTestCase {
     }
     
     func testUpdateDeviceSession() async throws {
-        try await seedCurrentUser()
+        try await service.seedCurrentUser()
         var tokenString = ""
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoMobile
             )
@@ -364,9 +354,9 @@ final class UserTests: XCTestCase {
             XCTAssertEqual(privateInfo.deviceSessions.count, 1)
             XCTAssertEqual(privateInfo.deviceSessions[0].deviceInfo.id, DeviceInfo.testInfoMobile.id)
         })
-        try app.test(.POST, "users/login",
-                     headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
-                     beforeRequest: { req in
+        try await asyncTest(.POST, "users/login",
+                            headers: .authWith(username: CurrentUser.username, password: CurrentUser.password),
+                            beforeRequest: { req in
             try req.content.encode(
                 DeviceInfo.testInfoDesktop
             )
@@ -376,9 +366,9 @@ final class UserTests: XCTestCase {
             XCTAssertEqual(privateInfo.deviceSessions.count, 2)
             tokenString = try XCTUnwrap(privateInfo.sessionForDeviceId(DeviceInfo.testInfoDesktop.id)).accessToken
         })
-        try app.test(.PUT, "users/me/device",
-                     headers: .authWith(token: tokenString),
-                     beforeRequest: { req in
+        try await asyncTest(.PUT, "users/me/device",
+                            headers: .authWith(token: tokenString),
+                            beforeRequest: { req in
             try req.content.encode(UpdateDeviceSessionRequest(deviceName: "My computer"))
         }, afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
@@ -391,71 +381,71 @@ final class UserTests: XCTestCase {
     }
     
     func testGetUser() async throws {
-        let (_, photo) = try await seedUserWithPhoto(name: "Test", username: "test")
-        try app.test(.GET, "users/1", headers: .none, afterResponse: { res in
+        let (_, photo) = try await service.seedUserWithPhoto(name: "Test", username: "test")
+        try await asyncTest(.GET, "users/1", headers: .none, afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
             let user = try res.content.decode(UserInfo.self)
             XCTAssertEqual(user.id, 1)
             XCTAssertEqual(user.username, "test")
             XCTAssertEqual(user.photos?.count, 1)
-            try photo.removeFiles()
+            try service.removeFiles(for: photo)
         })
     }
     
     func testSearchUsers() async throws {
-        try await seedUsers(count: 2, namePrefix: "Name", usernamePrefix: "user")
-        try await seedUser(name: "Demo 1", username: "test1")
-        let (_, photo) = try await seedUserWithPhoto(name: "Demo 2", username: "test2")
+        try await service.seedUsers(count: 2, namePrefix: "Name", usernamePrefix: "user")
+        try await service.seedUser(name: "Demo 1", username: "test1")
+        let (_, photo) = try await service.seedUserWithPhoto(name: "Demo 2", username: "test2")
         // Search by id=2
-        try app.test(.GET, "users?s=2", headers: .none, afterResponse: { res in
+        try await asyncTest(.GET, "users?s=2", headers: .none, afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
             let users = try res.content.decode([UserInfo].self)
             XCTAssertEqual(users.count, 1)
             XCTAssertEqual(users[0].id, 2)
         })
         // Search by username "u(se)r*"
-        try app.test(.GET, "users?s=se", headers: .none, afterResponse: { res in
+        try await asyncTest(.GET, "users?s=se", headers: .none, afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
             let names = try res.content.decode([UserInfo].self).compactMap { $0.name }
             XCTAssertEqual(names.sorted(), ["Name 1", "Name 2"])
         })
         // Search by name "D(em)o *"
-        try app.test(.GET, "users?s=em", headers: .none, afterResponse: { res in
+        try await asyncTest(.GET, "users?s=em", headers: .none, afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
             let users = try res.content.decode([UserInfo].self).sorted(by: { $0.id! < $1.id! })
             XCTAssertEqual(users.compactMap { $0.name }, ["Demo 1", "Demo 2"])
             XCTAssertEqual(users.first?.photos?.count, 0)
             XCTAssertEqual(users.last?.photos?.count, 1)
-            try photo.removeFiles()
+            try service.removeFiles(for: photo)
         })
     }
     
     func testAddAndDeletePhotoOfUser() async throws {
-        try await seedCurrentUser()
+        try await service.seedCurrentUser()
         let fileId = UUID()
         let fileName = fileId.uuidString
         let fileType = "test"
         
         // "Upload" all files before adding photo
-        let uploadPath = try makeFakeUpload(fileName: fileName + "." + fileType, fileSize: 1)
-        let previewPath = try makeFakeUpload(fileName: fileName + "-preview." + fileType, fileSize: 1)
+        let uploadPath = try service.makeFakeUpload(fileName: fileName + "." + fileType, fileSize: 1)
+        let previewPath = try service.makeFakeUpload(fileName: fileName + "-preview." + fileType, fileSize: 1)
         XCTAssertTrue(FileManager.default.fileExists(atPath: uploadPath))
         XCTAssertTrue(FileManager.default.fileExists(atPath: previewPath))
         
-        try app.test(.POST, "users/me/photos", headers: .none, beforeRequest: { req in
+        try await asyncTest(.POST, "users/me/photos", headers: .none, beforeRequest: { req in
             try req.content.encode(UpdateUserRequest(photo: MediaInfo(id: fileId, fileType: fileType, fileSize: 1)))
         }, afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
             let userInfo = try res.content.decode(UserInfo.self)
             XCTAssertEqual(userInfo.photos?.count, 1)
-            XCTAssertEqual(userInfo.photos?[0].fileExists, true)
-            XCTAssertEqual(userInfo.photos?[0].previewExists, true)
+            XCTAssertTrue(service.fileExists(for: userInfo.photos![0]))
+            XCTAssertTrue(service.previewExists(for: userInfo.photos![0]))
         })
         
-        try app.test(.DELETE, "users/me/photos/\(fileId)", headers: .none, afterResponse: { res in
+        try await asyncTest(.DELETE, "users/me/photos/\(fileId)", headers: .none, afterResponse: { res in
             XCTAssertEqual(res.status, .ok, res.body.string)
         })
-
+        
         // Check if all files were removed
         XCTAssertFalse(FileManager.default.fileExists(atPath: uploadPath))
         XCTAssertFalse(FileManager.default.fileExists(atPath: previewPath))
